chart file updated

+++ coa/dal/chart.php
@@ -32,25 +32,9 @@ class chart {
     {
         global $db;
 		
-        $Sql = 'INSTER INTO ' .$this->TableName. '(name, description)';
-        $Sql = $Sql . 'VALUES("' . $ChartName . '","' . $ChartDescription . '")';
-        try{
-        $query = $db->prepare($Sql);
-        $query->execute();
-                $result['msg'] = true;
-                $result['data'] = $query->fetchAll(PDO::FETCH_ASSOC);
-        }catch(PDOException $e){
-                $result['msg'] = false;
-                $result['data'] = $e->getMessage();
-        }
-		
-	return $result;
-        
-    }
-	
-	public function Update($ChartName,$ChartDescription,$ChartId){
-            global $db;
-            $Sql = 'UPDATE '.$this->TableName.' SET name = '.$ChartName.', description='.$ChartDescription.' WHERE id ='.$ChartId;
+        if($ChartName != "" && $ChartDescription != ""){
+            $Sql = 'INSTER INTO ' .$this->TableName. '(name, description)';
+            $Sql = $Sql . 'VALUES("' . $ChartName . '","' . $ChartDescription . '")';
             try{
             $query = $db->prepare($Sql);
             $query->execute();
@@ -60,6 +44,33 @@ class chart {
                     $result['msg'] = false;
                     $result['data'] = $e->getMessage();
             }
+        }else{
+            $result['msg'] = false;
+            $result['data'] = 'Invalid parameters';
+        }
+		
+	return $result;
+        
+    }
+	
+	public function Update($ChartName,$ChartDescription,$ChartId){
+            global $db;
+            
+            if($ChartName != "" && $ChartDescription != "" && $ChartId != ""){
+                $Sql = 'UPDATE '.$this->TableName.' SET name = '.$ChartName.', description='.$ChartDescription.' WHERE id ='.$ChartId;
+                try{
+                $query = $db->prepare($Sql);
+                $query->execute();
+                        $result['msg'] = true;
+                        $result['data'] = $query->fetchAll(PDO::FETCH_ASSOC);
+                }catch(PDOException $e){
+                        $result['msg'] = false;
+                        $result['data'] = $e->getMessage();
+                }
+            }else{
+                $result['msg'] = false;
+                $result['data'] = 'Invalid parameters';
+            }
 
             return $result;
 	}
@@ -67,15 +78,20 @@ class chart {
 	public function Delete_record($ChartId){
             global $db;
 		
-            $Sql = "DELETE FROM " . $this->TableName . " WHERE id =" . $ChartId;
-            try{
-            $query = $db->prepare($Sql);
-            $query->execute();
-                $result['msg'] = true;
-                $result['data'] = $query->fetchAll(PDO::FETCH_ASSOC);
-            }catch(PDOException $e){
-                $result['msg'] = false;
-                $result['data'] = $e->getMessage();
+            if($ChartId != ""){
+                $Sql = "DELETE FROM " . $this->TableName . " WHERE id =" . $ChartId;
+                try{
+                $query = $db->prepare($Sql);
+                $query->execute();
+                    $result['msg'] = true;
+                    $result['data'] = $query->fetchAll(PDO::FETCH_ASSOC);
+                }catch(PDOException $e){
+                    $result['msg'] = false;
+                    $result['data'] = $e->getMessage();
+                }
+            }else{
+                 $result['msg'] = false;
+                 $result['data'] = 'Invalid parameters';
             }
 
             return $result;
